refactor: TsunamiTimeline 系モデルを専用ファイルへ分離
