+++ app/lib/feature/tsunami/ui/components/tsunami_timeline_overlay.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 
 import 'package:eqmonitor/core/designsystem/design_system_build_context_x.dart';
+import 'package:eqmonitor/feature/tsunami/data/model/tsunami_playback_selection_state.dart';
 import 'package:eqmonitor/feature/tsunami/data/model/tsunami_telegram_with_state.dart';
 import 'package:eqmonitor/feature/tsunami/data/notifier/tsunami_playback_selection_notifier.dart';
 import 'package:eqmonitor/feature/tsunami/data/notifier/tsunami_telegrams_provider.dart';
