+++ app/test/feature/tsunami/tsunami_timeline_to_public_test.dart
@@ -1,5 +1,6 @@
 import 'package:eqmonitor/core/model/telegram/telegram_type.dart';
-import 'package:eqmonitor/feature/tsunami/data/model/timeline/tsunami_timeline.dart';
+import 'package:eqmonitor/feature/tsunami/data/logic/tracked_tsunami_timeline_to_public_mapper.dart';
+import 'package:eqmonitor/feature/tsunami/data/logic/tsunami_telegrams_response_to_tracked_timeline_builder.dart';
 import 'package:eqmonitor/feature/tsunami/data/model/tracking/tracked_offshore_station.dart';
 import 'package:eqmonitor/feature/tsunami/data/model/tracking/tracked_region.dart';
 import 'package:eqmonitor/feature/tsunami/data/model/tracking/tracked_region_station.dart';
